// src/lib/strapi/types/common/Media.ts
// Interface automatically generated by schemas-to-ts, adapted for Strapi v5's flat GraphQL response

import { MediaFormat } from "./MediaFormat";

export interface Media {
  id: number;
  documentId: string;
  name: string;
  alternativeText: string;
  caption: string;
  width: number;
  height: number;
  formats: { thumbnail: MediaFormat; small: MediaFormat; medium: MediaFormat; large: MediaFormat };
  hash: string;
  ext: string;
  mime: string;
  size: number;
  url: string;
  previewUrl: string;
  provider: string;
  createdAt: Date;
  updatedAt: Date;
}